Add EmptyPass, options in Ctx
This commit introduces a new type of pass, EmptyPass, which does
nothing. This pass can be used to replace a pass that has been disabled
(eg. `MainInjection` when `nativeSharedLibrary` is true).

Also, the `Ctx` gained an `options: Opts` field that contains the
options specific to scala-native. They make it easier to export
configuration to specific phases.