Simplify and extend implement of `java.io`
The previous implementation was an excellent base to get started. This
commits builds on this foundation and adds many of the missing classes,
while simplifying the implementation of others.

This commit also introduce numerous tests for the parts of `java.io`
that are re-implemented.

Using scripted tests for this may seem overkill, however it allows us to
depend on the JVM for setting up the tests, and makes it simple to
ensure that we correctly mimic the JVM's behavior.