Add `@stub` annotation to mark stubs (#972)
* Annotation to mark stubs

This commit adds an annotation `@stub` that can be added to method
definitions. Methods that have the `@stub` annotation will not link by
default (i.e. if something depends on it, it won't be resolved).

To allow `@stub` methods to be linked, one needs to set `linkStubs =
true` in `scala.scalanative.tools.Config`.

Fixes #971

* Tests for stubs

* Document `@stub` in nir documentation

* Annotate stubs from #968 with `@stub`

* Bump NIR revision

* Address review comments

* Add `nativeLinkStub` settingKey

This setting key controls whether methods that have the `@stub`
annotation should be included or ignored by the linker.

The default value is `false`, meaning that the linker will ignore
methods that have the `@stub` annotation.

* Remove `@stub` annotation for Enumeration

`scala.Enumeration` relies on reflection. We bypass the calls to the
reflection API, but they still exist and we must link with the stubs.
Thus we cannot give them the `@stub` annotation.

* Support `@stub` on classes and modules

* Address review comments