Fix ProcessSuite hangs (#1190)
* Fix stack corruption in release

At some point I switched from fork to vfork and while tests passed, I
never tested in release mode. I discovered while debugging the
ProcessSuite hangs that invoking a process always segfaults in release
mode on my mac. I believe this was because the child process modifies
the stack of the parent. The fix is to just create a new function that
runs the child code so that the child creates a new stack frame.

After this change, I ran a bash script that invoked a simple scala
native program that called git status and was compiled in release mode.
In over 1000 iterations of the script, I saw no crashes. (Prior to the
change, it always segfaulted)

* Fix undefined process behavior

It turns out that it's unsafe to signal back onto the main thread if the
main thread is performing gc. To work around this, I made it so that the
process results are cached in native code and the scala native process
must fetch them either by explicitly calling waitFor or by implicitly
calling any of a number of relevant apis. In particular, whenever the
user calls a function that checks the exit value, we check the native
cache and whenever we read or check the available bytes.

Unlike the old implementation, we will leak native file descriptors (and
an int -> int map entry) if the user never calls waitFor (or one of the
other functions that can have the implicit side effect of draining the input
streams and setting the proc result). I think this can be fixed by
caching the output and error streams in the native code as well, but I'm
going to defer that work.