Fix #1173: explicitly specify clang -std for C files (#1273)
Modified LLVM.scala to explicitly set -std=gnu11 when compiling
.c files.  Scala Native currently does not specify a standard
when compiling .c files.  Since at least Clang 3.7 this means
that the gnu11 standard is used.

Explicitly setting -std=gnu11 does not change the -std used
by Clang, but it does make the standard in use more evident
when nativeCompileOptions specifies the -v verbose flag.

The intent is to make it easier for contributors to Scala Native
to discover the C standard being used. The C++ standard of
C++11 has shown up in verbose output for years.

It is helpful for contributors, especially new contributors,
to know the C standard without having to do days of searching
Clang defaults because some symbols start with double underscore
(__) when compiled under -std=c11. They do not have this prefix
when compiled under -std=gnu11.  If I remember correctly,
timezone is one of these names.