Add support for DeleteOnExit (#1222)
This commit adds support for arbitrary shutdown hooks in scala native.
On the jvm there are a number of hooks that run when the program exits,
including the delete on exit hooks. I discovered that c++ will call the
destructors of static member variables on program exit. These static member
variables are also lazily initialized which ensures that the hooks are
only run if any are added. There may be a better approach to this, but
this does seem to work reliably for me. With the new native Hook class,
I was able to almost trivially add deleteOnExit to java.io.File.

I didn't add the application hooks that the jvm provides because I
wasn't sure if this approach would be acceptable to the community.