Add support for ThinLTO (#1281)
Previously we relied on one large LLVM file
for the whole application as our means to let
LLVM optimize whole program together.

This change adds supports for ThinLTO introduced
LLVM 3.9. It lets us compile the last step in parallel
and also offers 5.2% throughput improvement
due to inlining across Scala/C boundary which
our previous approach could not do.

Unfortunately we can not enable it by default due
to compatibility constraints. Recent versions of
clang on linux and ubuntu support it out of the box,
but it's not always available on older distributions
of linux.