Make unwind handler blocks first class (#1399)
Previously, unwind handlers were rather second class and slightly magical:

...
%y = call %f(%x) unwind %handle
...
%handle(%e):
...

Here %handle is a an exception handling block that may only be used as target for unwind jumps. Value for %e appears magically through mapping to LLVM's landingpads in codegen. This makes regular jumps to unwind handler blocks ill-behaved as they would hit a landing pad that's inserted in the block prologue.

Instead, this PR changes them to be more uniform with normal control-flow. In particular, unwind jumps now define a value (that corresponds to the caught exception) together with conventional parametrized jump. For example:

...
%y = call %f(%x) unwind %e to %handle(%e)
...
%handle(%e):
...

The handler block doesn't need to differentiate unwind and non-unwind incoming edges any more, it only cares about the regular parametrized jump with an exception as an argument. The landing pads are generated for each unwind definition. It's impossible to accidentally jump to landingpad in any way except for genuine unwind handling case.

This opens the door for exception handler blocks to be usable as targets for regular jumps in combination with exceptional unwind jumps at the same time. This is important to enable Interflow transform exceptional control-flow into regular control-flow (which is in turn is necessary to optimize non-local returns).