Interflow: initial infrastructure (#1339)
This change introduces preliminary infrastructure for Interflow (#1278)

Here, we focus on type inference and method duplication.
Partial evaluation and partial escape analysis are included as well
but they are mostly inactive until we add inlining.

We leave interflow enabled at all times on master to catch potential
bugs early. Due to added compile time cost, it's likely we'll make it
a release mode only feature in the future.